test: integration tests for users
